+++ app/admin/biz/handler/admin/role/role_service.go
@@ -144,7 +144,7 @@ func UpdateRoleStatus(ctx context.Context, c *app.RequestContext) {
 		return
 	}
 
-	err = admin.NewRole(ctx, c).UpdateStatus(int64(req.ID), int8(req.Status))
+	err = admin.NewRole(ctx, c).UpdateStatus(int64(req.ID), req.Status)
 	if err != nil {
 		utils.SendResponse(c, errno.ConvertErr(err), nil, 0, "")
 		return
